# [SPARK-32978][SQL] Make sure the number of dynamic part metric is correct