--- a/sql/core/src/test/scala/org/apache/spark/sql/execution/benchmark/InsertTableWithDynamicPartitionsBenchmark.scala
+++ b/sql/core/src/test/scala/org/apache/spark/sql/execution/benchmark/InsertTableWithDynamicPartitionsBenchmark.scala
@@ -37,7 +37,7 @@ object InsertTableWithDynamicPartitionsBenchmark extends DataSourceWriteBenchmar
     SparkSession.builder().master("local[4]").getOrCreate()
   }
 
-  def prepareSourceTableAndGetTotalRowsCount(numberRows: Long,
+  def prepareSourceTableAndGetTotalRows(numberRows: Long, sourceTable: String,
       part1Step: Int, part2Step: Int, part3Step: Int): Long = {
     val dataFrame = spark.range(0, numberRows, 1, 4)
     val dataFrame1 = spark.range(0, numberRows, part1Step, 4)
@@ -46,74 +46,48 @@ object InsertTableWithDynamicPartitionsBenchmark extends DataSourceWriteBenchmar
 
     val data = dataFrame.join(dataFrame1).join(dataFrame2).join(dataFrame3)
       .toDF("id", "part1", "part2", "part3")
-
-    data.createOrReplaceTempView("tmpTable")
-
-    spark.sql("create table " +
-      "sourceTable(id bigint, part1 bigint, part2 bigint, part3 bigint) " +
-      "using parquet")
-
-    spark.sql("insert overwrite table sourceTable " +
-      s"select id, " +
-      s"part1, " +
-      s"part2, " +
-      s"part3 " +
-      s"from tmpTable")
-
-    spark.catalog.dropTempView("tmpTable")
+    data.write.saveAsTable(sourceTable)
     data.count()
   }
 
-  def prepareTable(): Unit = {
-    spark.sql("create table " +
-      "tableOnePartitionColumn(i bigint, part bigint) " +
+  def writeOnePartitionColumnTable(tableName: String,
+      partitionNumber: Long, benchmark: Benchmark): Unit = {
+    spark.sql(s"create table $tableName(i bigint, part bigint) " +
       "using parquet partitioned by (part)")
-    spark.sql("create table " +
-      "tableTwoPartitionColumn(i bigint, part1 bigint, part2 bigint) " +
-      "using parquet partitioned by (part1, part2)")
-    spark.sql("create table " +
-      "tableThreePartitionColumn(i bigint, part1 bigint, part2 bigint, part3 bigint) " +
-      "using parquet partitioned by (part1, part2, part3)")
-  }
-
-  def writeOnePartitionColumnTable(partitionNumber: Long, benchmark: Benchmark): Unit = {
     benchmark.addCase(s"one partition column, $partitionNumber partitions") { _ =>
       spark.sql("insert overwrite table " +
-        "tableOnePartitionColumn partition(part) " +
-        s"select id, " +
-        s"part1 as part " +
-        s"from sourceTable")
+        s"$tableName partition(part) " +
+        "select id, part1 as part from sourceTable")
     }
   }
 
-  def writeTwoPartitionColumnTable(partitionNumber: Long, benchmark: Benchmark): Unit = {
+  def writeTwoPartitionColumnTable(tableName: String,
+      partitionNumber: Long, benchmark: Benchmark): Unit = {
+    spark.sql(s"create table $tableName(i bigint, part1 bigint, part2 bigint) " +
+      "using parquet partitioned by (part1, part2)")
     benchmark.addCase(s"two partition columns, $partitionNumber partitions") { _ =>
       spark.sql("insert overwrite table " +
-        "tableTwoPartitionColumn partition(part1, part2) " +
-        s"select id, " +
-        s"part1, " +
-        s"part2 " +
-        s"from sourceTable")
+        s"$tableName partition(part1, part2) " +
+        "select id, part1, part2 from sourceTable")
     }
   }
 
-  def writeThreePartitionColumnTable(partitionNumber: Long, benchmark: Benchmark): Unit = {
+  def writeThreePartitionColumnTable(tableName: String,
+      partitionNumber: Long, benchmark: Benchmark): Unit = {
+    spark.sql(s"create table $tableName(i bigint, part1 bigint, part2 bigint, part3 bigint) " +
+      "using parquet partitioned by (part1, part2, part3)")
     benchmark.addCase(s"three partition columns, $partitionNumber partitions") { _ =>
       spark.sql("insert overwrite table " +
-        "tableThreePartitionColumn partition(part1, part2, part3) " +
-        s"select id, " +
-        s"part1, " +
-        s"part2, " +
-        s"part3 " +
-        s"from sourceTable")
+        s"$tableName partition(part1, part2, part3) " +
+        "select id, part1, part2, part3 from sourceTable")
     }
   }
 
   override def runBenchmarkSuite(mainArgs: Array[String]): Unit = {
     val sourceTable = "sourceTable"
-    val tableOnePartitionColumn = "tableOnePartitionColumn"
-    val tableTwoPartitionColumn = "tableTwoPartitionColumn"
-    val tableThreePartitionColumn = "tableThreePartitionColumn"
+    val onePartColTable = "onePartColTable"
+    val twoPartColTable = "twoPartColTable"
+    val threePartColTable = "threePartColTable"
     val numberRows = 100L
     val part1Step = 1
     val part2Step = 20
@@ -121,17 +95,16 @@ object InsertTableWithDynamicPartitionsBenchmark extends DataSourceWriteBenchmar
     val part1Number = numberRows / part1Step
     val part2Number = numberRows / part2Step *  part1Number
     val part3Number = numberRows / part3Step *  part2Number
-    withTable(sourceTable, tableOnePartitionColumn, tableTwoPartitionColumn,
-      tableThreePartitionColumn) {
+
+    withTable(sourceTable, onePartColTable, twoPartColTable, threePartColTable) {
       val totalRows =
-        prepareSourceTableAndGetTotalRowsCount(numberRows, part1Step, part2Step, part3Step)
-      prepareTable()
+        prepareSourceTableAndGetTotalRows(numberRows, sourceTable, part1Step, part2Step, part3Step)
       val benchmark =
         new Benchmark(s"dynamic insert table benchmark, totalRows = $totalRows",
           totalRows, output = output)
-      writeOnePartitionColumnTable(part1Number, benchmark)
-      writeTwoPartitionColumnTable(part2Number, benchmark)
-      writeThreePartitionColumnTable(part3Number, benchmark)
+      writeOnePartitionColumnTable(onePartColTable, part1Number, benchmark)
+      writeTwoPartitionColumnTable(twoPartColTable, part2Number, benchmark)
+      writeThreePartitionColumnTable(threePartColTable, part3Number, benchmark)
       benchmark.run()
     }
   }