@LuciferYang 2020-10-13
cc @wangyum , the current result should be correct, but will increase memory and network pressure