@AngersZhuuuu 2020-10-13
I have thought to do like this too, but if partition num is too big, return partition value will occupy too many memory.